+++ transektcount/src/main/java/com/wmstein/transektcount/widgets/CountingWidgetHead1.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 -2023. Wilhelm Stein, Bonn, Germany.
+ * Copyright © 2016-2025. Wilhelm Stein, Bonn, Germany.
  */
 package com.wmstein.transektcount.widgets
 
@@ -18,27 +18,25 @@ import com.wmstein.transektcount.R
  * Created by wmstein 18.12.2016,
  * last edited in Java on 2023-05-09,
  * converted to Kotlin on 2023-06-26,
- * last edited on 2023-10-06
+ * last edited on 2025-07-16
  */
 class CountingWidgetHead1(
     context: Context,
-    resource: Int,
-    private val idArray: Array<String>,       // species ids
-    private val contentArray1: Array<String>, // species names
-    private val contentArray3: Array<String>, // species local names
-    private val contentArray2: Array<String>, // species codes
-    private val imageArray: Array<Int>        // species images
-) : ArrayAdapter<String?>(context, resource, R.id.countName, contentArray1) {
-    private val inflater: LayoutInflater
+    private val idArray: Array<String>,    // species ids
+    private val nameArray: Array<String>,  // species names
+    private val nameArrayL: Array<String>, // species local names
+    private val codeArray: Array<String>,  // species codes
+    private val imageArray: Array<Int>     // species images
+) : ArrayAdapter<String?>(context, R.layout.widget_counting_head1, R.id.countName, nameArray) {
 
-    init {
-        inflater = context.getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
-    }
+    private val inflater: LayoutInflater = context.getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
 
+    // Shows Spinner list
     override fun getDropDownView(position: Int, convertView: View?, parent: ViewGroup): View {
         return getCustomView(position, parent)
     }
 
+    // Shows rest of counting page
     override fun getView(position: Int, convertView: View?, parent: ViewGroup): View {
         return getCustomView(position, parent)
     }
@@ -48,13 +46,14 @@ class CountingWidgetHead1(
         val countId = head1.findViewById<TextView>(R.id.countId) // used for spinner interface  
         countId.text = idArray[position]
         val countName = head1.findViewById<TextView>(R.id.countName)
-        countName.text = contentArray1[position]
+        countName.text = nameArray[position]
         val countNameg = head1.findViewById<TextView>(R.id.countNameg)
-        countNameg.text = contentArray3[position]
+        countNameg.text = nameArrayL[position]
         val countCode = head1.findViewById<TextView>(R.id.countCode)
-        countCode.text = contentArray2[position]
+        countCode.text = codeArray[position]
         val pSpecies = head1.findViewById<ImageView>(R.id.pSpecies)
         pSpecies.setImageResource(imageArray[position])
         return head1
     }
+
 }