+++ transektcount/src/main/java/com/wmstein/transektcount/widgets/CountingWidgetInt.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 - 2025. Wilhelm Stein, Bonn, Germany.
+ * Copyright © 2016-2025, Wilhelm Stein, Bonn, Germany.
  */
 package com.wmstein.transektcount.widgets